Consolidate & fix the RequestStatus component
The actual props didn't match the Flow types, which opened up a rabbit hole of tracking down dead code. The code that remains is just the code that actually ran.
