--- src/modules/UI/components/RequestStatus/index.js
@@ -1,5 +1,0 @@
-/* eslint-disable flowtype/require-valid-file-annotation */
-
-import RequestStatus from './RequestStatus.ui.js'
-
-export default RequestStatus
